Problems concerned
The package of code functions VC en- and de-cryption properly given txt file. 
fileProcess class turn txt file to string, ready for further process. 
Note that in String instance, each "\n" occupies two units in storage which is not the same as it is processed in the tutorial, it however functions well.